refactor(backend): move auth tests into a separate file
